benchmark: [BACKUP] qdefender to prevent microbenchmark optimizations
Just a way to prevent compiler optimizations by removing operations whose results are not consumed later.

Comparing to the previous code, this version adds up time in the decapsulate, but: 1-expected because in real life we consume the decapsulate, and 2- the difference is not big, so regarding decapsulate time, we can discard that time.

Nevertheless, this commit is just a backup because for gcc 9 on linux, the previous code and this code returns similar times. Since this bench is not intended for strict timing comparisons, we will revert changes to keep the benchmark test as simple as possible.

---- Kyber1024 -----
KEYGEN: 0.020000 ms/op
ENCAPS: 0.025000 ms/op
DECAPS: 0.018000 ms/op

----- NTRU-HPS-4096-821 -----
KEYGEN: 0.230000 ms/op
ENCAPS: 0.141000 ms/op
DECAPS: 0.019000 ms/op

----- NTRU-HRSS-1373 -----
KEYGEN: 4.670000 ms/op
ENCAPS: 0.148000 ms/op
DECAPS: 0.178000 ms/op