benchmark: qdefender, go back to the lean version of benchmark
Note that for other environments, architectures, the previous benchmark code could make sense. However, since most of the test for this bench are performed on pc/laptops with x86 arch and gcc (no clang or other compilers) better to keep the code clean and we will go back to microptimizations when needed.